fix: window 객체 필요에 따른 ssr 비활성화
